requirements.txt
nltk~=3.5
numpy~=1.18.5
Keras~=2.4.3
PyYAML~=5.3.1
Flask~=1.1.2
tensorflow
